application/radar_host/requirements.txt
PySide6>=6.7,<7.0
pyqtgraph>=0.13.7,<0.14
pyserial>=3.5,<4.0
